requirements.txt
anthropic>=0.40.0
playwright>=1.44.0
playwright-stealth>=1.0.6
beautifulsoup4>=4.12.0
lxml>=5.2.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.127.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
httpx>=0.27.0
